fix: move the component file
